move browser utils to dedicated module
